Use docker image with home directory tar
When the user launches the container, a lot of time is spent on preparing the environment, setting up codes, pseudopotentials, etc. In order to reduce the launch time of the container, we tar already prepared home directory and keep it in the image, and then untar it when the container is launched. This way, the user can start working with the container immediately.